+++ src/main/java/com/hhs/codeboard/blog/data/entity/board/dto/BoardCategoryDto.java
@@ -14,6 +14,6 @@ public class BoardCategoryDto {
 
     private YN useF;
 
-    private Integer boardSeq;
+    private Long boardSeq;
 
 }
