+++ src/main/java/com/hhs/codeboard/blog/data/entity/menu/dto/MenuDto.java
@@ -19,7 +19,7 @@ public class MenuDto extends DefaultSearchDto {
 
     private Integer menuOrder;
 
-    private Integer parentSeq;
+    private Long parentSeq;
 
     private String uuid;
